backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
requests
PyJWT
pydantic-settings
psycopg2-binary
